Week 14: Reading Sound Files and Analyzing Pure Tones
The current code combines all functionality in one class (the
PureToneTest class). Here a pure tone can be generated based on user
input or by hard-coded values.  Although the generation based on user
input has been commented out in the makePureTone method it can easily be
substituted in.

The an array of byte values are written based on the array of double
values that represent pure tones. These byte values are written to a
.WAV file and stored in the same directory. These steps are taken for
testing purposes; to ensure the portion of the program that attempts to
read the sound file operates appropriately.
This portion is critical for analyzing actual notes and stepping forward
from just pure tones. Once it attempts to read the bytes that represent
the sound file, it converts them to doubles for frequency analysis.

The frequency analysis seems to work (although somewhat inefficiently)
for pure tones yet not for actual notes.

Timing values have also been extracted and displayed using a sliding
window. That is, for every time window of a preset time a frequency
analysis is performed to determine changes in detected frequency values.
When a change is noted, a time stamp is stored and attached to the new
perceived note. This method is only effective for pure tones that differ
from the ones played before and after them.

The testPureTone.wav and testPureTone2.wav files are the sound files
that were analyzed for testing purposes.

testRecord.wav contains the sound recorded from the microphone while
testing it out.

The files 01.wav, eqt-major-sc.wav and just-major-sc.wav contain sound
from actual instruments being played and will be used for testing out
the program in later stages.