17th March Update
Changed from DFT to FFT. Cuts runtime from 88 seconds to 2 seconds for
test file.
Quicker frequency analysis allows for transition to real-time processing
(should be included in the next update).

Things to work on (in decreasing order of perceived difficulty):
Timing analysis - determining tempo and time signatures from the sounds
heard.
Transcription - There were delays with this update because of issues
with allowing chords to be written. If no progress is made by the end of
the month a new transcription library may be necessary.
Real-time processing - samples will be taken from microphone and
directly processed as opposed to storing them as sound files and doing
post-recording analyses.